fix(auth_core): 修改返回参数类型为long

+++ src/main/java/cn/rainier/nian/dao/MenuDao.java
@@ -97,6 +97,6 @@ public interface MenuDao{
 	
 	public List<Menu> findMenuByRole(String roleName);
 	
-	public int findChildMenuCount(Integer menuId);
+	public long findChildMenuCount(Integer menuId);
 }
